+++ src/main/java/ru/vgoudk/workhours/repositories/PositionRepository.java
@@ -3,5 +3,5 @@ package ru.vgoudk.workhours.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.vgoudk.workhours.model.finance.Position;
 
-public interface PositionRepository extends JpaRepository<Position, Long> {
+public interface PositionRepository extends JpaRepository<Position, Integer> {
 }
